# openapi_server/models/random_article_with_derivation.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class RandomArticleWithDerivation(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    RandomArticleWithDerivation - a model defined in OpenAPI

        article: The article of this RandomArticleWithDerivation [Optional].
        derivation: The derivation of this RandomArticleWithDerivation [Optional].
    """

    article: Optional[int] = Field(alias="article", default=None)
    derivation: Optional[List[str]] = Field(alias="derivation", default=None)

RandomArticleWithDerivation.update_forward_refs()